Added sync-exams command to pull unpulled exams.
See TIGRLab/mritool#24. This commit adds 'sync-exams' which keeps a log file of the pulled
exam ids in logs/exams.txt. Any exams on the server not recorded in this file are pulled
when this command is run. There isn't very much in the way of error-checking at the moment,
because I'm not sure what to anticipate.