Changed shutil.move to .copyfile to avoid copying file metadata
See https://github.com/TIGRLab/mritool/issues/32. shutil.move attempts to copy the file and the metadata
for a file to the destination. This doesn't work if the file is owned by someone else (even if the
group permissions allow writing to the file). shutil.copyfile just copies the file contents.