Fix for check/list-inprocess failing on pdfs.
Fixes TIGRLab/mritool#52

The issue is that a non-dicom file in an exam folder would cause a stack because
we weren't properly testing that a path was a directory before descending.

This fix also includes an optimization that stops indexing an exam dicom once a
single dicom file has been found in the exam folder (rather than indexing *every*
dicom file).